+++ configs/configs.go
@@ -138,7 +138,13 @@ func getOrSetDefaults(envVar string) {
 		}
 	} else {
 		s := viper.GetString(envVar)
-		os.Setenv(envVar, EnsureTrailingSlash(s))
+
+		// envs that need a trailing slash
+		if envVar == "GHORG_SCM_TYPE" || envVar == "GHORG_ABSOLUTE_PATH_TO_CLONE_TO" {
+			os.Setenv(envVar, EnsureTrailingSlash(s))
+		} else {
+			os.Setenv(envVar, s)
+		}
 	}
 }